Separate out XML processing tests for clarity
